+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -242,6 +242,18 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "networkTimeout",
+        name = "Base Network Timeout",
+        description = "The maximum number of seconds permitted to connect/read the webhook URL upon notifications",
+        position = 1002,
+        section = advancedSection
+    )
+    @Units(Units.SECONDS)
+    default int networkTimeout() {
+        return 15; // elevated from okhttp default of 10
+    }
+
+    @ConfigItem(
         keyName = "imageWriteTimeout",
         name = "Image Upload Timeout",
         description = "The maximum number of seconds that uploading a screenshot can take before timing out",
@@ -551,6 +563,30 @@ public interface DinkPluginConfig extends Config {
     )
     default String customPlayerBadge() {
         return "";
+    }
+
+    @ConfigItem(
+        keyName = "proxyServer",
+        name = "Proxy Server",
+        description = "HTTPS proxy server for posting notifications; must be in host:port format.<br/>" +
+            "You may need to restart the client for changes to this setting to take effect",
+        position = 1024,
+        section = advancedSection
+    )
+    default String proxyServer() {
+        return ""; // host:port
+    }
+
+    @ConfigItem(
+        keyName = "proxyAuth",
+        name = "Proxy Auth",
+        description = "Authentication for the above proxy server in user:pass format.<br/>" +
+            "You may need to restart the client for changes to this setting to take effect",
+        position = 1025,
+        section = advancedSection
+    )
+    default String proxyAuth() {
+        return ""; // user:pass
     }
 
     @ConfigItem(