+++ src/main/java/dinkplugin/message/DiscordMessageHandler.java
@@ -8,6 +8,8 @@ import dinkplugin.domain.SeasonalPolicy;
 import dinkplugin.message.templating.Replacements;
 import dinkplugin.message.templating.Template;
 import dinkplugin.notifiers.data.NotificationData;
+import dinkplugin.util.ConfigProxyAuth;
+import dinkplugin.util.ConfigProxyServer;
 import dinkplugin.util.ConfigUtil;
 import dinkplugin.util.DiscordProfile;
 import dinkplugin.util.Utils;
@@ -99,7 +101,9 @@ public class DiscordMessageHandler {
                 Request request = chain.request().newBuilder()
                     .header("User-Agent", DinkPlugin.USER_AGENT)
                     .build();
-                Interceptor.Chain updatedChain = chain;
+                Interceptor.Chain updatedChain = chain
+                    .withConnectTimeout(config.networkTimeout(), TimeUnit.SECONDS)
+                    .withReadTimeout(config.networkTimeout(), TimeUnit.SECONDS);
                 // Allow longer timeout when writing a screenshot file to overcome slow internet speeds
                 if (request.body() instanceof MultipartBody && Utils.hasImage((MultipartBody) request.body())) {
                     updatedChain = chain.withWriteTimeout(Math.max(config.imageWriteTimeout(), 0), TimeUnit.SECONDS);
@@ -125,6 +129,8 @@ public class DiscordMessageHandler {
                 }
                 return response;
             })
+            .proxySelector(new ConfigProxyServer(config))
+            .proxyAuthenticator(new ConfigProxyAuth(config))
             .build();
     }